+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [feature] date parser tool: add timestamp support.
 * [feature] date parser tool: add helper text.
 * [feature] base64: add image picker.
+* [feature] add password generator tool.
 * [fix] number base: set input limit.
 * [fix] fix copy error.
 * [fix] update windows make config.